refactor #2: 공통 예외 디렉토리 수정 및 변수명 통일
