--- src/main/java/com/wanted/spendtracker/domain/ErrorCode.java
@@ -1,19 +1,0 @@
-package com.wanted.spendtracker.domain;
-
-import static org.springframework.http.HttpStatus.*;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@AllArgsConstructor
-@Getter
-public enum ErrorCode {
-
-    COMMON_INVALID_PARAMETER("잘못된 파라미터입니다.", BAD_REQUEST),
-    COMMON_SERVER_ERROR("서버에서 에러가 발생하였습니다.", INTERNAL_SERVER_ERROR)
-    ;
-
-    private final String message;
-    private final HttpStatus httpStatus;
-}
