+++ src/main/java/com/wanted/spendtracker/exception/CustomException.java
@@ -1,4 +1,4 @@
-package com.wanted.spendtracker.domain;
+package com.wanted.spendtracker.exception;
 
 import lombok.Getter;
